Move PyStringType helper methods inside PyStringType; Add docs
